requirements.txt
Flask==1.1.4
Pillow==6.2.2
Numpy==1.16.4
opencv-python==4.3.0.38



